Passing 126 with Mocha specs. Currently troubleshooting Bonus Bookings routes

+++ backend/db/models/spot.js
@@ -56,7 +56,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     lat: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.FLOAT,
       // unique: true,
       allowNull: false, 
       validate: {
@@ -65,7 +65,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     lng: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.FLOAT,
       // unique: true,
       allowNull: false,
       validate: {
